Extract host details and auto generate schema
