+++ .gitignore
@@ -1,3 +1,6 @@
 .env
 /.venv
 /.vscode
+/timescaledb
+__pycache__
+ref
