+++ Dockerfile.exporter
@@ -1,5 +1,12 @@
 FROM python:3.8
 
+RUN apt-get update
+RUN apt-get install -y apt-transport-https ca-certificates curl gnupg-agent software-properties-common
+RUN curl -fsSL https://download.docker.com/linux/debian/gpg | apt-key add -
+RUN add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/debian $(lsb_release -cs) stable"
+RUN apt-get update
+RUN apt-get install -y docker-ce-cli
+
 RUN mkdir /app
 COPY ./requirements.txt /app
 
@@ -8,4 +15,4 @@ RUN pip install -r requirements.txt
 
 COPY /src /app
 
-CMD python docker_host_exporter.py
+CMD python host_exporter.py
